Normalization of barcode paramaters.
The idea is that end-users don't not need to know specificities of each
ESC/POS implementation to be able to print barcodes. They will rely on a
default set of parameter values and the implementations will take care of
the details, trying to use best values based on the given parameters.

Also, doctests are now enabled through `setup.cfg`.