Tests directory moved out of escpos package
The idea of using pipenv was abandoned. Abstract dependencies
will be listed in setup.py via install_requires, as well as
optional dependencies via extras_require.

The "requirements" directory is there for development purposes
only, as requirements/dev.txt file points to an editable
dependency to the main 'escpos' package.

* Now pytest is invoked via tox or via py.test;
* Pinned pytest >= 4.6 and < 5 for Python 2 and 3 compatibility;
* Pipenv specific files where removed;
* Tox file describes many scenarios from Python 2.7 to 3.8 including
  scenarios for optional depencencies;
* Improved manifest file for correct source distribution.