Backoff retry exception handling changed
Previously, a list of exceptions indicate when to retry in case of an
exception is thrown. Now we pass a callable which is responsible for
handling exceptions and decide to retry or not. This allows for a more
fine grained control over exception handling and the backoff retry
algorithm.

Added automatic handling for some type of exceptions that is common
during network write operations under very unstable connections. The
following errors will cause a disconnection before delay and a
reconnection before retry (on write only):

- EPIPE (broken pipe)
- ENOTCONN (transport endpoint not connected)
- ECONNRESET (connection reset by peer)